android: Move service client into it's own file
The service file was getting harder to read so lets pull the client code
into it's own file
