+++ android/app/src/main/java/org/musicpd/AutomationReceiver.kt
@@ -1,0 +1,9 @@
+package org.musicpd
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+
+class AutomationReceiver : BroadcastReceiver() {
+    override fun onReceive(context: Context, intent: Intent) {}
+}
